Moved all logic from the `TestRunner` out into a controller file
