+++ Jantimizer/ExperimentSuite/UserControls/TestRunner/Controls/HistogramGeneratorControl.xaml.cs
@@ -26,28 +26,9 @@ namespace ExperimentSuite.UserControls
             InitializeComponent();
         }
 
-        public async Task GenerateHistograms(IHistogramManager manager)
+        private void HistogramProgressBar_ValueChanged(object sender, RoutedPropertyChangedEventArgs<double> e)
         {
-            List<Task> tasks = await manager.AddHistogramsFromDB();
-            HistogramProgressBar.Maximum = tasks.Count;
-            HistogramProgressBar.Value = 0;
-            Update_HistogramProgressLabel(0, (int)HistogramProgressBar.Maximum);
-            // https://docs.microsoft.com/en-us/dotnet/csharp/programming-guide/concepts/async/start-multiple-async-tasks-and-process-them-as-they-complete?pivots=dotnet-6-0#create-the-asynchronous-sum-page-sizes-method
-            while (tasks.Any())
-            {
-                var finishedTask = await Task.WhenAny(tasks);
-                tasks.Remove(finishedTask);
-                await finishedTask;
-                HistogramProgressBar.Value++;
-                Update_HistogramProgressLabel((int)HistogramProgressBar.Value, (int)HistogramProgressBar.Maximum);
-            }
-            HistogramProgressBar.Value = HistogramProgressBar.Maximum;
-            Update_HistogramProgressLabel((int)HistogramProgressBar.Maximum, (int)HistogramProgressBar.Maximum);
-        }
-
-        private void Update_HistogramProgressLabel(int current, int max)
-        {
-            HistogramProgressLabel.Content = $"Histogram Progress ({current}/{max})";
+            HistogramProgressLabel.Content = $"Histogram Progress ({HistogramProgressBar.Value}/{HistogramProgressBar.Maximum})";
         }
     }
 }
