+++ Jantimizer/ExperimentSuite/UserControls/TestRunner/Controls/SQLFileProgressControl.xaml
@@ -8,7 +8,6 @@
              d:DesignHeight="450" d:DesignWidth="800">
     <StackPanel>
         <Label x:Name="SQLProgressLabel" Content="SQL Progress (0/0)"/>
-        <Label x:Name="CurrentSqlFileLabels" Content="File: " FontSize="10"/>
         <ProgressBar x:Name="SQLProgressBar" Height="20" ValueChanged="SQLProgressBar_ValueChanged"/>
     </StackPanel>
 </UserControl>
